+++ lib/instream/response.ex
@@ -7,6 +7,8 @@ defmodule Instream.Response do
   Maybe parses a response based on the requested result type.
   """
   @spec maybe_parse(response :: any, opts :: Keyword.t) :: any
+  def maybe_parse("", _), do: nil
+
   def maybe_parse(response, opts) do
     case opts[:result_as] do
       :raw -> response