+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enhancements
   - Allows accessing raw query results (undecoded binaries) using `Instream.Connection.execute/2`
   - Read queries (binaries) can be executed on a database
+  - Write queries (map data) can be executed on a database
 
 ## v0.1.0 (2015-02-23)